update due cifre decimali

+++ Ristorante/src/main/java/model/Pagamento.java
@@ -1,13 +1,15 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Pagamento {
 	private int id_pagamento;
 	private int id_tavolo;
-	private float costo_totale;
+	private BigDecimal costo_totale;
 	private String stato;
 	private String data;
 
-	public Pagamento(int id,int id_tavolo, float ct, String s, String d) {
+	public Pagamento(int id,int id_tavolo, BigDecimal ct, String s, String d) {
 		this.id_pagamento=id;
 		this.id_tavolo = id_tavolo;
 		this.costo_totale = ct;
@@ -34,11 +36,11 @@ public class Pagamento {
 		this.id_tavolo = id_tavolo;
 	}
 
-	public float getCosto_totale() {
+	public BigDecimal getCosto_totale() {
 		return costo_totale;
 	}
 
-	public void setCosto_totale(float costo_totale) {
+	public void setCosto_totale(BigDecimal costo_totale) {
 		this.costo_totale = costo_totale;
 	}
